services/resume-parser/requirements.txt
flask==3.1.0
flask-cors==5.0.0
supabase==2.11.0
openai==1.59.9
cffi==1.17.0
pdfminer.six==20240706
structlog==24.4.0
requests==2.32.3
ruff==0.8.3
pre-commit==4.0.1
python-docx==1.1.2
pytz==2024.2
werkzeug==3.1.3
jinja2==3.1.4
pillow==10.4.0
mammoth==1.8.0
PyMuPDF==1.24.12
numpy==1.26.4
opencv-python-headless==4.10.0.84
python-dotenv==1.0.1
gunicorn==23.0.0